+++ plib/resources/locales/en-US.php
@@ -1,6 +1,9 @@
 <?php
 // Copyright 1999-2016. Parallels IP Holdings GmbH.
 $messages = array(
+    // Pathbar
+    'controllers.index.index.title' => 'Slave DNS Manager',
+
     'listPageTitle' => 'Slave DNS Manager',
     'statusColumnTitle' => 'Status',
     'configColumnTitle' => 'IP Address',
@@ -10,15 +13,21 @@ $messages = array(
     'refreshToolDescription' => 'Refresh remote slaves status',
     'removeToolTitle' => 'Remove',
     'removeToolDescription' => 'Remove the selected slave configurations.',
+    'resyncToolTitle' => 'Resync',
+    'resyncToolDescription' => 'Resync all DNS zones from master to slaves.',
 
     'addPageTitle' => 'Creating Remote Slave Configuration',
     'howTo' => 'Add to named.conf on the remote server (config location depends on the OS of the slave server)',
     'learnMore' => 'Learn more about the slave server setup.',
-    'ipLabel' => 'IP address',
+    'masterIpLabel' => 'Plesk (master) IP address',
+    'ipLabel' => 'Slave DNS IP address',
+    'invalidIpExistingSlave' => 'This Slave DNS already exist',
     'portLabel' => 'Port',
     'algorithmLabel' => 'Algorithm',
     'secretLabel' => 'Secret',
     'invalidSecret' => 'The secret should be a base64 encoded string.',
+    'viewLabel' => 'View',
+    'invalidAlnumDashUnderscore' => 'The view should consists of letters, digits, dashes, underscores.',
     'slaveSaved' => 'The slave configuration was saved.',
 
     'viewPageTitle' => 'Remote Slave Configuration',
@@ -26,7 +35,9 @@ $messages = array(
 
     'emptySelection' => 'Select items first.',
     'removeConfirmation' => 'Remove the selected configurations?',
-    'removeConfirmationYes' => 'Yes',
-    'removeConfirmationNo' => 'No',
+    'resyncConfirmation' => 'Resync all DNS zones from master to slaves?',
+    'confirmationYes' => 'Yes',
+    'confirmationNo' => 'No',
+    'resyncDone' => 'All DNS zones were synced from master to slaves.',
     'slaveRemoved' => 'The slave configurations were removed.',
 );