+++ CHANGES.md
@@ -1,9 +1,17 @@
+# 1.8.0
+* [+] Add support choose server's IP (issue [#13](https://github.com/plesk/ext-slave-dns-manager/issues/13))
+* [-] Fixed incompatibility with rndc on RedHat (issue [#8](https://github.com/plesk/ext-slave-dns-manager/issues/8))
+* [+] Add support Split DNS (issue [#3](https://github.com/plesk/ext-slave-dns-manager/issues/3), [#7](https://github.com/plesk/ext-slave-dns-manager/issues/7))
+* [*] The extension saves settings in key-value storage
+* [*] Small increase verbosity in status icon tooltip if error occurred
+* [+] Add button 'Resync' to sync all DNS zones from master to slaves
+* [+] Add navigation pathbar
 
-# 1.7.2
+# 1.7.2 (4 April 2017)
 * [+] Add "Troubleshooting" in README.md
-* [*] Change link for [[learnMore]] from "http://devblog.plesk.com" to "https://www.plesk.cloud"
+* [*] Change link for [[learnMore]] from "http://devblog.plesk.com" to "https://www.plesk.com"
 * [+] Add RU description in meta.xml
-* [-] Fix category in meta.xml
+* [-] Fixed category in meta.xml
 * [+] Add DESCRIPTION.md
 * [+] Add CHANGES.md